+++ generate_dataset/generate_natural_language/__init__.py
@@ -1,1 +1,2 @@
-from .generate_nl import generate_nl, CustomDataset, Example
+from .generate_nl import generate_nl
+from .dataset_class import Example, CustomDataset
