+++ generate_dataset/parse_funcs/__init__.py
@@ -1,2 +1,1 @@
 from .parse_derivation import parse_derivations
-from .base_classes import BaseIndividual, BaseOperator, Term, Assertion, Formula
