+++ generate_dataset/rules/conic10k.cfg
@@ -1,0 +1,112 @@
+# 在top数据集中，不存在TERM = TERM，只有TERM。当然严谨来说，这里应该称作一个formula / fact / proposition等
+$START=$Assertion
+# 这里收录所有的算子。每个FACT是由predicate (xx)组成的。但考虑到CFG的表达能力，我们这里需要罗列predicate+对应arity的组合
+# 而不是分别用PREDICATE和ARITY两个非终止节点代替
+$Assertion=Coordinate ( $POINT ) = ( $NUMBER $NUMBER )
+|XCoordinate ( $POINT ) = ( $NUMBER )
+|YCoordinate ( $POINT ) = ( $NUMBER )
+|LocusEquation ( $POINT ) = ( $EXPRESSION )
+|Locus ( $POINT ) = ( $CURVE )
+|Quadrant ( $POINT ) = ( $NUMBER )
+|Focus ( $CONICSECTION ) = ( $POINTS )
+|RightFocus ( $CONICSECTION ) = ( $POINT )
+|LeftFocus ( $CONICSECTION ) = ( $POINT )
+|LowerFocus ( $CONICSECTION ) = ( $POINT )
+|UpperFocus ( $CONICSECTION ) = ( $POINT )
+|FocalLength ( $CONICSECTION ) = ( $NUMBER )
+|HalfFocalLength ( $CONICSECTION ) = ( $NUMBER )
+|SymmetryAxis ( $CONICSECTION ) = ( $AXIS )
+|Eccentricity ( $CONICSECTION ) = ( $NUMBER )
+|Vertex ( $CONICSECTION ) = ( $POINTS )
+|UpperVertex ( $ELLIPSE_HYPERBOLA ) = ( $POINT )
+|LowerVertex ( $ELLIPSE_HYPERBOLA ) = ( $POINT )
+|LeftVertex ( $ELLIPSE_HYPERBOLA ) = ( $POINT )
+|RightVertex ( $ELLIPSE_HYPERBOLA ) = ( $POINT )
+|Directrix ( $CONICSECTION ) = ( $LINES )
+|LeftDirectrix ( $CONICSECTION ) = ( $LINE )
+|RightDirectrix ( $CONICSECTION ) = ( $LINE )
+|MajorAxis ( $ELLIPSE ) = ( $LINESEGMENT )
+|MinorAxis ( $ELLIPSE ) = ( $LINESEGMENT )
+|RealAxis ( $ELLIPSE ) = ( $LINESEGMENT )
+|ImageinaryAxis ( $ELLIPSE ) = ( $LINESEGMENT )
+|LeftPart ( $HYPERBOLA ) = ( $CURVE )
+|RightPart ( $HYPERBOLA ) = ( $CURVE )
+|Asymptote ( $HYPERBOLA ) = ( $TWOLINES )
+|Diameter ( $DIAMETER ) = ( $NUMBER )
+|Radius ( $DIAMETER ) = ( $NUMBER )
+|Perimeter ( $CIRCLE_ELLIPSE ) = ( $NUMBER )
+|Center ( $CIRCLE_ELLIPSE ) = ( $POINT )
+|Expression( $CURVE ) = ( $EXPRESSION )
+|LineOf( $POINT $POINT ) = ( $LINE )
+|Slope( $LINE ) = ( $NUMBER )
+|Inclination( $LINE ) = ( $NUMBER )
+|Intercept( $LINE $AXIS ) = ( $NUMBER )
+|LineSegmentOf( $POINT $POINT ) = ( $LINESEGMENT )
+|Length( $LINESEGMENT ) = ( $NUMBER )
+|MidPoint( $LINESEGMENT ) = ( $POINT )
+|OverlappingLine( $LINESEGMENT ) = ( $LINE )
+|PerpendicularBisector( $LINESEGMENT ) = ( $LINE )
+|Endpoint( $LINESEGMENT ) = ( $TWOPOINTS )
+|Endpoint( $RAY ) = ( $POINT )
+|Projection( $POINT $LINE ) = ( $POINT )
+|Projection( $POINT $AXIS ) = ( $POINT )
+|InterceptChord( $LINE $CONICSECTION ) = ( $LINESEGMENT )
+|TriangleOf( $POINT $POINT $POINT ) = ( $TRIANGLE )
+|InscribedCircle( $TRIANGLE ) = ( $CIRCLE )
+|CircumCircle( $TRIANGLE ) = ( $CIRCLE )
+|Incenter( $TRIANGLE ) = ( $POINT )
+|Orthocenter( $TRIANGLE ) = ( $POINT )
+|Circumcenter( $TRIANGLE ) = ( $POINT )
+|AngleOf( $POINT $POINT $POINT ) = ( $NUMBER )
+|VectorOf( $POINT $POINT ) = ( $VECTOR )
+|DotProduct( $VECTOR $VECTOR ) = ( $NUMBER )
+|NumIntersection( $AXIS_CURVE $AXIS_CURVE ) = ( $NUMBER )
+|Intersection( $AXIS_CURVE $AXIS_CURVE ) = ( $POINTS )
+|FootPoint( $LINE_LINESEGMENT $LINE_LINESEGMENT ) = ( $POINT )
+|DISTANCE( $POINT_LINE $POINT_LINE ) = ( $NUMBER )
+|Area( $CIRCLE ) = ( $NUMBER )
+|Area( $ELLIPSE ) = ( $NUMBER )
+|Area( $TRIANGLE ) = ( $NUMBER )
+|TangentPoint( $LINE $CIRCLE ) = ( $POINT )
+|TangentOfPoint( $POINT $CURVE ) = ( $LINE )
+|TangentOnPoint( $POINT $CURVE ) = ( $LINE )
+|Abs( $NUMBER ) = ( $NUMBER )
+|Abs( $LINESEGMENT ) = ( $NUMBER )
+|lg( $NUMBER ) = ( $NUMBER )
+|InterReciprocal( $NUMBER ) = ( $NUMBER )
+|sqrt( $NUMBER ) = ( $NUMBER )
+|Max( $NUMBER ) = ( $NUMBER )
+|Min( $NUMBER ) = ( $NUMBER )
+|Range( $NUMBER ) = ( $NUMBER $NUMBER )
+|WhenMin( $NUMBER )
+|WhenMax( $NUMBER )
+|ApplyUnit( $NUMBER $DEGREE )
+|ApplyUnit( $NUMBER $RAD )
+|In( $NUMBER ( $NUMBER $NUMBER ) )
+|Between( $POINT $POINT $POINT )
+|IsIntersect( $AXIS_CURVE $AXIS_CURVE )
+|IsPerpendicular( $LINE_LINESEGMENT_AXIS $LINE_LINESEGMENT_AXIS )
+|IsParallel( $LINE_LINESEGMENT_AXIS $LINE_LINESEGMENT_AXIS )
+|IsDiameter( $LINESEGMENT $CIRCLE )
+|IsChordOf( $LINESEGMENT $CONICSECTION )
+|IsTangent( $LINE $CONICSECTION )
+|IsInscribedCircle( $CIRCLE $TRIANGLE )
+|IsCircumCircle( $CIRCLE $TRIANGLE )
+|IsInTangent( $CIRCLE $CIRCLE )
+|IsOutTangent( $CIRCLE $CIRCLE )
+|PointOnCurve( $POINT $AXIS_CURVE )
+|And( $Assertion $Assertion )
+|Negation( $Assertion )
+
+$POINT = [ POINT: A-Z ]
+$CONICSECTION = [ CONICSECTION: a-zA-Z ]
+$CIRCLE = [ CIRCLE: a-zA-Z ]
+$NUMBER = [ NUMBER: number_generate ]
+$EXPRESSION = [ EXPRESSION: expression_generate ]
+$CURVE = [ CURVE: curve_generate ]
+$POINTS = [ POINTS: POINT ]
+|[ POINTS:  ]   # 多个点的集合
+$AXIS = [ AXIS: xAxis]
+|[ AXIS: yAxis]
+$AXIS_CURVE = [ AXIS_CURVE: AXIS]
+|[ AXIS_CURVE: CURVE ]
